# server/api/openapi.yaml
openapi: 3.1.0
info:
  title: Ghost Shift HTTP API
  version: 0.1.0
  description: Public, internal, operational, and observability endpoints exposed by Ghost Shift.
servers:
  - url: /
paths:
...
  /api/public/analytics/trends:
    get:
      summary: Get trend changes across the past N hours
      operationId: getAnalyticsTrends
      parameters:
        - $ref: "#/components/parameters/Hours"
        - $ref: "#/components/parameters/Since"
        - $ref: "#/components/parameters/Until"
      responses:
        "200":
          description: Trend response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicTrendResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /api/public/zones/heatmap:
    get:
      summary: Get zone activity heatmap
      operationId: getZonesHeatmap
      parameters:
        - $ref: "#/components/parameters/Since"
        - $ref: "#/components/parameters/Until"
      responses:
        "200":
          description: Zone heatmap
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicZonesHeatmapResponse"
  /api/public/models/distribution:
    get:
      summary: Get model distribution metrics
      operationId: getModelsDistribution
      parameters:
        - $ref: "#/components/parameters/Since"
        - $ref: "#/components/parameters/Until"
      responses:
        "200":
          description: Model distribution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicModelsDistributionResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
  parameters:
    Hours:
      name: hours
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
    Since:
      name: since
      in: query
      required: false
      schema:
        type: string
        format: date-time
    Until:
      name: until
      in: query
      required: false
      schema:
        type: string
        format: date-time
  responses:
    BadRequest:
      description: Invalid query parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorBody:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
      required: [code, message]
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
        requestId:
          type: string
      required: [error]
...
    PublicStatusShare:
      type: object
      properties:
        status:
          type: string
        count:
          type: integer
        share:
          type: number
      required: [status, count, share]
...
    PublicZonesHeatmapResponse:
      type: object
      properties:
        capturedAt:
          type: string
          format: date-time
        zones:
          type: array
          items:
            type: object
            properties:
              zone:
                type: string
              activityScore:
                type: number
              agentCount:
                type: integer
              statusDistribution:
                type: array
                items:
                  $ref: "#/components/schemas/PublicStatusShare"
            required: [zone, activityScore, agentCount, statusDistribution]
      required: [capturedAt, zones]
    PublicModelsDistributionResponse:
      type: object
      properties:
        models:
          type: array
          items:
            type: object
            properties:
              model:
                type: string
              share:
                type: number
              sampleCount:
                type: integer
              agentCount:
                type: integer
              avgResponseTime:
                type: number
              throughputTokensPerMinute:
                type: number
              avgLoad:
                type: number
            required: [model, share, sampleCount, agentCount, avgResponseTime, throughputTokensPerMinute, avgLoad]
      required: [models]
...
    PublicTrendPoint:
      type: object
      properties:
        capturedAt:
          type: string
          format: date-time
        onlineAgents:
          type: integer
        runningAgents:
          type: integer
        messageCount:
          type: integer
        totalTokens:
          type: integer
        deltaMessages:
          type: integer
        deltaToolCalls:
          type: integer
        deltaTokens:
          type: integer
        avgResponseTime:
          type: number
      required: [capturedAt, onlineAgents, runningAgents, messageCount, totalTokens, deltaMessages, deltaToolCalls, deltaTokens, avgResponseTime]
    PublicTrendSummary:
      type: object
      properties:
        sampleCount:
          type: integer
        onlineAgentsDelta:
          type: integer
        runningAgentsDelta:
          type: integer
        messageCountDelta:
          type: integer
        toolCallCount:
          type: integer
        totalTokensDelta:
          type: integer
        avgResponseTime:
          type: number
      required: [sampleCount, onlineAgentsDelta, runningAgentsDelta, messageCountDelta, toolCallCount, totalTokensDelta, avgResponseTime]
    PublicTrendResponse:
      type: object
      properties:
        hours:
          type: integer
        since:
          type: string
          format: date-time
        until:
          type: string
          format: date-time
        points:
          type: array
          items:
            $ref: "#/components/schemas/PublicTrendPoint"
        summary:
          $ref: "#/components/schemas/PublicTrendSummary"
      required: [hours, since, until, points, summary]
...